refactor: reply dto 구조 리팩토링 (#95)
