+++ application/src/main/java/com/foodielog/application/reply/controller/ReplyController.java
@@ -1,7 +1,8 @@
 package com.foodielog.application.reply.controller;
 
-import com.foodielog.application.reply.dto.ReplyCreatDTO;
-import com.foodielog.application.reply.dto.ReportReplyDTO;
+import com.foodielog.application.reply.dto.request.ReplyCreatReq;
+import com.foodielog.application.reply.dto.request.ReportReplyReq;
+import com.foodielog.application.reply.dto.response.ReplyCreatResp;
 import com.foodielog.application.reply.service.ReplyService;
 import com.foodielog.server._core.security.auth.PrincipalDetails;
 import com.foodielog.server._core.util.ApiUtils;
@@ -24,12 +25,12 @@ public class ReplyController {
     private final ReplyService replyService;
 
     @PostMapping("/{feedId}")
-    public ResponseEntity<ApiUtils.ApiResult<ReplyCreatDTO.Response>> saveReply(
+    public ResponseEntity<ApiUtils.ApiResult<ReplyCreatResp>> saveReply(
             @AuthenticationPrincipal PrincipalDetails principalDetails,
             @PathVariable Long feedId,
-            @Valid @RequestBody ReplyCreatDTO.Request createDTO,
+            @Valid @RequestBody ReplyCreatReq createDTO,
             Errors errors) {
-        ReplyCreatDTO.Response response = replyService.createReply(principalDetails.getUser(), feedId, createDTO);
+        ReplyCreatResp response = replyService.createReply(principalDetails.getUser(), feedId, createDTO);
         return new ResponseEntity<>(ApiUtils.success(response, HttpStatus.CREATED), HttpStatus.CREATED);
     }
 
@@ -43,19 +44,19 @@ public class ReplyController {
     }
 
     @GetMapping("/{feedId}")
-    public ResponseEntity<ApiUtils.ApiResult<ReplyCreatDTO.ListDTO>> getReplyList(
+    public ResponseEntity<ApiUtils.ApiResult<ReplyCreatResp.ListDTO>> getReplyList(
             @PathVariable Long feedId,
             @RequestParam Long replyId,
             @PageableDefault Pageable pageable
     ) {
-        ReplyCreatDTO.ListDTO response = replyService.getListReply(feedId, replyId, pageable);
+        ReplyCreatResp.ListDTO response = replyService.getListReply(feedId, replyId, pageable);
         return new ResponseEntity<>(ApiUtils.success(response, HttpStatus.OK), HttpStatus.OK);
     }
 
     @PostMapping("/report")
     public ResponseEntity<ApiUtils.ApiResult<String>> report(
             @AuthenticationPrincipal PrincipalDetails principalDetails,
-            @RequestBody @Valid ReportReplyDTO.Request request,
+            @RequestBody @Valid ReportReplyReq request,
             Errors errors
     ) {
         replyService.reportReply(principalDetails.getUser(), request);
